move file, ready for further metrics work
